pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Daniel Espinoza <danielx.1997@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.13"
streamlit = "^1.32.2"
asyncio = "^3.4.3"
pymupdf = "^1.24.1"
tqdm = "^4.66.2"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
semantic-kernel = "0.9.5b1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

